Move DnaDef et al to zome_types, with feature flag
