+++ crates/holochain/src/core/ribosome/guest_callback.rs
@@ -12,7 +12,7 @@ use crate::core::ribosome::FnComponents;
 use crate::core::ribosome::Invocation;
 use crate::core::ribosome::RibosomeT;
 use fallible_iterator::FallibleIterator;
-use holochain_types::dna::zome::Zome;
+use holochain_types::prelude::*;
 use holochain_zome_types::ExternIO;
 
 pub struct CallIterator<R: RibosomeT, I: Invocation> {
@@ -79,7 +79,6 @@ mod tests {
     use crate::fixt::ZomeCallHostAccessFixturator;
     use crate::fixt::ZomeFixturator;
     use fallible_iterator::FallibleIterator;
-    use holochain_types::dna::zome::Zome;
     use holochain_types::prelude::*;
     use mockall::predicate::*;
     use mockall::Sequence;
