+++ crates/holochain/src/core/ribosome/guest_callback/entry_defs.rs
@@ -4,7 +4,6 @@ use crate::core::ribosome::Invocation;
 use crate::core::ribosome::ZomesToInvoke;
 use derive_more::Constructor;
 use holochain_serialized_bytes::prelude::*;
-use holochain_types::dna::zome::HostFnAccess;
 use holochain_types::prelude::*;
 use std::collections::BTreeMap;
 
@@ -91,7 +90,7 @@ mod test {
     use crate::fixt::EntryDefsInvocationFixturator;
     use crate::fixt::ZomeNameFixturator;
     use ::fixt::prelude::*;
-    use holochain_types::dna::zome::HostFnAccess;
+    use holochain_types::prelude::*;
     use holochain_zome_types::entry_def::EntryDefsCallbackResult;
     use holochain_zome_types::ExternIO;
     use std::collections::BTreeMap;
