+++ crates/holochain/src/core/ribosome/guest_callback/post_commit.rs
@@ -96,7 +96,7 @@ mod test {
     use crate::fixt::PostCommitHostAccessFixturator;
     use crate::fixt::PostCommitInvocationFixturator;
     use ::fixt::prelude::*;
-    use holochain_types::dna::zome::HostFnAccess;
+    use holochain_types::prelude::*;
     use holochain_zome_types::post_commit::PostCommitCallbackResult;
     use holochain_zome_types::ExternIO;
