+++ crates/holochain/src/core/ribosome/guest_callback/validation_package.rs
@@ -122,7 +122,7 @@ mod test {
     use crate::core::ribosome::ZomesToInvoke;
     use crate::fixt::ValidationPackageHostAccessFixturator;
     use crate::fixt::ValidationPackageInvocationFixturator;
-    use holochain_types::dna::zome::HostFnAccess;
+    use holochain_types::prelude::*;
     use holochain_zome_types::validate::ValidationPackage;
     use holochain_zome_types::validate::ValidationPackageCallbackResult;
     use holochain_zome_types::ExternIO;
@@ -170,7 +170,7 @@ mod test {
 
     #[tokio::test(flavor = "multi_thread")]
     async fn validation_package_invocation_allow_side_effects() {
-        use holochain_types::dna::zome::Permission::*;
+        use holochain_types::access::Permission::*;
         let validation_package_host_access =
             ValidationPackageHostAccessFixturator::new(::fixt::Unpredictable)
                 .next()
