+++ crates/holochain/src/sweettest/sweet_dna.rs
@@ -1,12 +1,4 @@
-use holochain_types::{
-    dna::{
-        error::DnaResult,
-        random_uid, wasm,
-        zome::{inline_zome::InlineZome, Zome, ZomeDef},
-        DnaDefBuilder, DnaFile,
-    },
-    prelude::DnaDef,
-};
+use holochain_types::prelude::*;
 use holochain_zome_types::zome::ZomeName;
 use std::path::Path;
