+++ crates/holochain_types/src/dna.rs
@@ -4,16 +4,13 @@
 //! as well as serializing and deserializing dna, mainly to json format.
 
 mod dna_bundle;
-mod dna_def;
 mod dna_file;
 mod dna_manifest;
 mod dna_store;
 
 pub mod error;
 pub mod wasm;
-pub mod zome;
 pub use dna_bundle::*;
-pub use dna_def::*;
 pub use dna_file::*;
 pub use dna_manifest::*;
 pub use dna_store::MockDnaStore;
