+++ crates/holochain_types/src/dna/dna_bundle.rs
@@ -1,6 +1,5 @@
 use std::{collections::BTreeMap, path::PathBuf};
 
-use super::{DnaDef, DnaFile, WasmMap};
 use crate::prelude::*;
 use holo_hash::*;
 use mr_bundle::Location;
@@ -197,7 +196,6 @@ mod tests {
     use std::path::PathBuf;
 
     use super::*;
-    use crate::prelude::ZomeManifest;
 
     #[tokio::test(flavor = "multi_thread")]
     async fn dna_bundle_to_dna_file() {
