+++ crates/holochain_types/src/dna/error.rs
@@ -1,16 +1,12 @@
 //! Holochain DnaError type.
 
 use holo_hash::{DnaHash, WasmHash};
-use holochain_zome_types::zome::ZomeName;
+use holochain_zome_types::zome::error::ZomeError;
 use thiserror::Error;
 
 /// Holochain DnaError type.
 #[derive(Debug, Error)]
 pub enum DnaError {
-    /// ZomeNotFound
-    #[error("Zome not found: {0}")]
-    ZomeNotFound(String),
-
     /// EmptyZome
     #[error("Zome has no code: {0}")]
     EmptyZome(String),
@@ -39,6 +35,10 @@ pub enum DnaError {
     #[error(transparent)]
     SerializedBytesError(#[from] holochain_serialized_bytes::SerializedBytesError),
 
+    /// From ZomeError
+    #[error(transparent)]
+    ZomeError(#[from] ZomeError),
+
     /// std::io::Error
     /// we don't #[from] the std::io::Error directly because it doesn't implement Clone
     #[error("std::io::Error: {0}")]
@@ -47,10 +47,6 @@ pub enum DnaError {
     /// InvalidWasmHash
     #[error("InvalidWasmHash")]
     InvalidWasmHash,
-
-    /// NonWasmZome
-    #[error("Accessed a zome expecting to find a WasmZome, but found other type. Zome name: {0}")]
-    NonWasmZome(ZomeName),
 
     /// DnaHashMismatch
     #[error("DNA file hash mismatch.\nExpected: {0}\nActual: {1}")]
