+++ crates/holochain_types/src/fixt.rs
@@ -2,15 +2,9 @@
 
 #![allow(missing_docs)]
 
-use crate::dna::zome::HostFnAccess;
-use crate::dna::zome::Permission;
-use crate::dna::zome::Zome;
-use crate::dna::zome::ZomeDef;
-use crate::dna::DnaDef;
-use crate::dna::Zomes;
 use crate::header::NewEntryHeader;
+use crate::prelude::*;
 use ::fixt::prelude::*;
-use holochain_zome_types::prelude::*;
 use rand::seq::IteratorRandom;
 use std::iter::Iterator;
