+++ crates/holochain_types/src/lib.rs
@@ -12,6 +12,7 @@
 // We have a lot of usages of type aliases to `&String`, which clippy objects to.
 #![allow(clippy::ptr_arg)]
 
+pub mod access;
 pub mod activity;
 pub mod app;
 pub mod autonomic;
@@ -27,6 +28,7 @@ pub mod link;
 mod macros;
 pub mod metadata;
 pub mod prelude;
+pub mod properties;
 pub mod signal;
 pub mod timestamp;
 pub mod validate;
