+++ crates/holochain_types/src/prelude.rs
@@ -8,6 +8,7 @@ pub use holochain_zome_types::prelude::*;
 pub use std::convert::TryFrom;
 pub use std::convert::TryInto;
 
+pub use crate::access::*;
 pub use crate::activity::*;
 pub use crate::app::error::*;
 pub use crate::app::*;
@@ -18,9 +19,6 @@ pub use crate::dht_op::error::*;
 pub use crate::dht_op::*;
 pub use crate::dna::error::*;
 pub use crate::dna::wasm::*;
-pub use crate::dna::zome::inline_zome::error::*;
-pub use crate::dna::zome::inline_zome::*;
-pub use crate::dna::zome::*;
 pub use crate::dna::*;
 pub use crate::element::error::*;
 pub use crate::element::*;
@@ -29,8 +27,9 @@ pub use crate::header::error::*;
 pub use crate::header::*;
 pub use crate::link::*;
 pub use crate::metadata::*;
+pub use crate::properties::*;
 pub use crate::signal::*;
-pub use crate::timestamp; // for timestmap::now()
+pub use crate::timestamp; // for timestamp::now()
 pub use crate::timestamp::*;
 pub use crate::validate::*;
