+++ crates/holochain_zome_types/Cargo.toml
@@ -27,9 +27,18 @@ fixt = { version = "^0.0.2-alpha.1", path = "../fixt", optional = true }
 strum = { version = "0.18.0", optional = true }
 rand = {version = "0.7", optional = true}
 
+# full-dna-def dependencies
+derive_builder = { version = "0.9", optional = true }
+nanoid = { version = "0.3", optional = true }
+shrinkwraprs = { version = "0.3", optional = true }
+
 [features]
-default = ["fixturators", "test_utils"]
+default = ["full-dna-def", "fixturators", "test_utils"]
+
+# Extra impls for DnaDef, including InlineZome, which are not used in Wasm
+# but used in Holochain
+full-dna-def = ["derive_builder", "nanoid", "shrinkwraprs"]
 
 fixturators = ["fixt", "rand", "strum", "holo_hash/fixturators", ]
 
-test_utils = ["holo_hash/hashing", "holo_hash/test_utils"]
+test_utils = ["holo_hash/hashing", "holo_hash/test_utils", "full-dna-def"]
