+++ crates/holochain_zome_types/src/dna_def.rs
@@ -1,56 +1,14 @@
-use super::error::DnaError;
+//! Defines DnaDef struct
+
 use super::zome;
-use crate::prelude::*;
+use crate::{prelude::*, zome::error::ZomeError};
 use holo_hash::*;
-use holochain_zome_types::ZomeName;
 
 /// Zomes need to be an ordered map from ZomeName to a Zome
 pub type Zomes = Vec<(ZomeName, zome::ZomeDef)>;
 
-/// A type to allow json values to be used as [SerializedBytes]
-#[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    Eq,
-    derive_more::From,
-    serde::Serialize,
-    serde::Deserialize,
-    SerializedBytes,
-)]
-pub struct YamlProperties(serde_yaml::Value);
-
-impl YamlProperties {
-    /// Create new properties from json value
-    pub fn new(properties: serde_yaml::Value) -> Self {
-        Self(properties)
-    }
-
-    /// Create a null set of properties
-    pub fn empty() -> Self {
-        Self(serde_yaml::Value::Null)
-    }
-}
-
-impl From<()> for YamlProperties {
-    fn from(_: ()) -> Self {
-        Self::empty()
-    }
-}
-
-impl Default for YamlProperties {
-    fn default() -> Self {
-        Self::empty()
-    }
-}
-
-/// Not a great implementation: always returns null
-#[cfg(feature = "arbitrary")]
-impl<'a> arbitrary::Arbitrary<'a> for YamlProperties {
-    fn arbitrary(_: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
-        Ok(serde_yaml::Value::Null.into())
-    }
-}
+/// Placeholder for a real UID type
+pub type Uid = String;
 
 /// The definition of a DNA: the hash of this data is what produces the DnaHash.
 ///
@@ -61,13 +19,15 @@ impl<'a> arbitrary::Arbitrary<'a> for YamlProperties {
 /// TODO: after removing the `InstallApp` admin method, we can remove the Serialize
 ///       impl on this type, and document it/rename it to show that it is
 ///       basically a fully validated, normalized DnaManifest
-#[derive(
-    Serialize, Deserialize, Clone, Debug, PartialEq, Eq, SerializedBytes, derive_builder::Builder,
-)]
-#[builder(public)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, SerializedBytes)]
+#[cfg_attr(feature = "full-dna-def", derive(derive_builder::Builder))]
+#[cfg_attr(feature = "full-dna-def", builder(public))]
 pub struct DnaDef {
     /// The friendly "name" of a Holochain DNA.
-    #[builder(default = "\"Generated DnaDef\".to_string()")]
+    #[cfg_attr(
+        feature = "full-dna-def",
+        builder(default = "\"Generated DnaDef\".to_string()")
+    )]
     pub name: String,
 
     /// A UID for uniquifying this Dna.
@@ -75,7 +35,7 @@ pub struct DnaDef {
     pub uid: String,
 
     /// Any arbitrary application properties can be included in this object.
-    #[builder(default = "().try_into().unwrap()")]
+    #[cfg_attr(feature = "full-dna-def", builder(default = "().try_into().unwrap()"))]
     pub properties: SerializedBytes,
 
     /// An array of zomes associated with your DNA.
@@ -95,48 +55,52 @@ impl DnaDef {
     }
 }
 
+#[cfg(feature = "full-dna-def")]
 impl DnaDef {
     /// Return a Zome
-    pub fn get_zome(&self, zome_name: &ZomeName) -> Result<zome::Zome, DnaError> {
+    pub fn get_zome(&self, zome_name: &ZomeName) -> Result<zome::Zome, ZomeError> {
         self.zomes
             .iter()
             .find(|(name, _)| name == zome_name)
             .cloned()
             .map(|(name, def)| Zome::new(name, def))
-            .ok_or_else(|| DnaError::ZomeNotFound(format!("Zome '{}' not found", &zome_name,)))
+            .ok_or_else(|| ZomeError::ZomeNotFound(format!("Zome '{}' not found", &zome_name,)))
     }
 
     /// Return a Zome, error if not a WasmZome
-    pub fn get_wasm_zome(&self, zome_name: &ZomeName) -> Result<&zome::WasmZome, DnaError> {
+    pub fn get_wasm_zome(&self, zome_name: &ZomeName) -> Result<&zome::WasmZome, ZomeError> {
         self.zomes
             .iter()
             .find(|(name, _)| name == zome_name)
             .map(|(_, def)| def)
-            .ok_or_else(|| DnaError::ZomeNotFound(format!("Zome '{}' not found", &zome_name,)))
+            .ok_or_else(|| ZomeError::ZomeNotFound(format!("Zome '{}' not found", &zome_name,)))
             .and_then(|def| {
                 if let ZomeDef::Wasm(wasm_zome) = def {
                     Ok(wasm_zome)
                 } else {
-                    Err(DnaError::NonWasmZome(zome_name.clone()))
+                    Err(ZomeError::NonWasmZome(zome_name.clone()))
                 }
             })
     }
 
     /// Change the "phenotype" of this DNA -- the UID and properties -- while
     /// leaving the "genotype" of actual DNA code intact
-    pub fn modify_phenotype(&self, uid: Uid, properties: YamlProperties) -> DnaResult<Self> {
+    #[cfg(feature = "full-dna-def")]
+    pub fn modify_phenotype(&self, uid: Uid, properties: SerializedBytes) -> Self {
         let mut clone = self.clone();
-        clone.properties = properties.try_into()?;
+        clone.properties = properties;
         clone.uid = uid;
-        Ok(clone)
+        clone
     }
 }
 
 /// Get a random UID
+#[cfg(feature = "full-dna-def")]
 pub fn random_uid() -> String {
     nanoid::nanoid!()
 }
 
+#[cfg(feature = "full-dna-def")]
 impl DnaDefBuilder {
     /// Provide a random UID
     pub fn random_uid(&mut self) -> &mut Self {
@@ -146,6 +110,7 @@ impl DnaDefBuilder {
 }
 
 /// A DnaDef paired with its DnaHash
+#[cfg(feature = "full-dna-def")]
 pub type DnaDefHashed = HoloHashed<DnaDef>;
 
 impl_hashable_content!(DnaDef, Dna);
