+++ crates/holochain_zome_types/src/prelude.rs
@@ -7,6 +7,7 @@ pub use crate::call_remote::*;
 pub use crate::capability::*;
 pub use crate::cell::*;
 pub use crate::crdt::*;
+pub use crate::dna_def::*;
 pub use crate::element::*;
 pub use crate::entry::*;
 pub use crate::entry_def::*;
@@ -29,6 +30,7 @@ pub use crate::trace::*;
 pub use crate::validate::*;
 pub use crate::validate_link::*;
 pub use crate::warrant::*;
+pub use crate::zome::error::*;
 pub use crate::zome::*;
 pub use crate::zome_io::ExternIO;
 pub use crate::zome_io::*;
@@ -39,7 +41,13 @@ pub use x_salsa20_poly1305::key_ref::*;
 pub use x_salsa20_poly1305::x25519::*;
 pub use x_salsa20_poly1305::*;
 
-// #[cfg(feature = "fixturators")]
+#[cfg(feature = "full-dna-def")]
+pub use crate::zome::inline_zome::error::*;
+#[cfg(feature = "full-dna-def")]
+pub use crate::zome::inline_zome::*;
+
+#[cfg(feature = "fixturators")]
 pub use crate::fixt::*;
+
 #[cfg(feature = "test_utils")]
 pub use crate::test_utils::*;
